Created info table
The info table in the CPEngLog database is intended to assist with verifying that the file is a CPEngLog file and with versioning.  Exisitange of and info table with a CPEngLogVer column verifies a CPEngLog DB.  The table is intended to hold a single row with the semantic version of the CPEngLog that was used to create the database.  This enables detection of old databases by future versions, and should facillitae migration.