Implemented check on opening and importing
Currently only valid sqlite3 databases initialise for CPEngLog are valid for opening.  Only excel files that are sourced the result of an export from the Engineers Australia myCPD database are valid.  The justification for reading the excel file into a pandas.DataFrame and then inserting to the sqlite database is that this separates the interpretation of excel and interaction with the database.  This separation facilitates reading of other file formats in future versions, without having to rework database interaction.  It could also facilitate interaction with other DBMS in future versions.